requirements.txt
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
requests>=2.28.0
pyyaml>=6.0
aiohttp>=3.8.0